[release] Changed install method and script to work correctly.
The problem with using "setup.py install" is that it builds a binary file for
the whole project. Since we use directory traversing to discover the modules, it
raises an error saying that the directory doesn't exist because it really
doesn't! So we have to install the files separately.

Pip allows us to install the whole project with its individual files and should
be the standard installation method. The main script was changed slightly to
comply with setup paths.